+++ CHANGELOG.md
@@ -30,7 +30,7 @@
       - [x] Enhancements in Overall pipeline execution of feature 'Google Takeout Processing'. Execution time reduced more than **50%** thanks to above enhancements.
       - [x] Enhancements in Steps execution order and logger messages. Now is clearer to follow the process pipeline.
     - [x] Enhancements in `build.py` module to reduce the Anti-Virus warning probability on Windows systems.
-    - [x] Renamed flag `-exeExifTool, --exec-exif-tool` by `-fnExtDates, --foldername-extracted_dates` to be more specific in the content of that output folder.
+    - [x] Renamed flag `-exeExifTool, --exec-exif-tool` by `-fnExtDates, --foldername-extracted-dates` to be more specific in the content of that output folder.
     
   - #### 🐛 Bug fixes:
     - [x] Fixed minor issues in Logger module. 
@@ -103,7 +103,7 @@
     - [x] Added new argument `-fnNoAlbums, --foldername-no/albums` to Allow user to define folder name for 'No-Albums'.
     - [x] Added new argument `-fnLogs, --foldername-logs` to Allow user to define folder name for 'Logs'.
     - [x] Added new argument `-fnDuplicat, --foldername-duplicates-output` to Allow user to define folder name for 'Duplicates Outputs'.
-    - [x] Added new argument `-fnExtDates, --foldername-extracted_dates` to Allow user to define folder name for 'Exiftool Outputs'.
+    - [x] Added new argument `-fnExtDates, --foldername-extracted-dates` to Allow user to define folder name for 'Exiftool Outputs'.
     - [x] Added new argument `-exeGpthTool, --exec-gpth-tool` to Allow user to specify an external version of GPTH Tool binary.
     - [x] Added new argument `-exeExifTool, --exec-exif-tool` to Allow user to specify an external version of EXIF Tool binary.
     - [x] Added new argument `-gSkipPrep, --google-skip-preprocess` to Skipp Preprocess steps during Google Takeout Processing feature.